TestBD/serializers.py
from rest_framework import serializers
from .models import Students

class StudentsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Students
        fields = ["id",'student_name', 'student_surname', 'student_age']